main

+++ FinAd/Controllers/newClientsData.cs
@@ -3,11 +3,11 @@
     public class newClientsData
     {
         public string? Name { get; set; }
-        public Int32? PhoneNo { get; set; }
+        public Int64? PhoneNo { get; set; }
         public string? Email { get; set; }
         
         public string? PancardNo { get; set; }
-        public Int32? AadharNo { get; set; }
+        public Int64? AadharNo { get; set; }
         public string? RiskProfile { get; set; }
         public string? SuggestedModel { get; set; }
        /* public string? AdvisorEmail { get; set; }*/
